pyproject.toml
[project]
name = "discord-issue-bot"
version = "0.1.0"
description = "Discord bot that generates GitHub issues from channel messages using Gemini AI."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "discord.py>=2.4",
    "google-genai>=1.0",
    "httpx>=0.27",
    "PyJWT[crypto]>=2.0",
    "python-json-logger>=3.0",

    "sentry-sdk>=2.0",

    "PyYAML>=6.0",

]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "prompt_eval: tests that call a real LLM API (requires GEMINI_API_KEY)",
]
